Simplify: deduplicate conversation extraction, fix N+1 fetch, use shared components
- Extract shared extractConversation() utility into utils/conversation.ts,
  removing duplicate logic from GroupDetailPage and RolloutDetailPage
- Fix N+1 fetch in GroupDetailPage: use getGroupRollouts() endpoint
  instead of fetching summary + N individual detail requests
- Use MetaField component in GroupDetailPage instead of inline styles
- Use rewardColor() from shared.tsx in RewardHistogram instead of
  duplicate binColor()
- Remove unnecessary ConvMessage type alias in RolloutDetailPage
